+++ src/HMR/Library.fs
@@ -1,7 +1,0 @@
-namespace Elmish.Bridge.HMR
-open Elmish.HMR
-open Elmish
-[<RequireQualifiedAccess>]
-module Bridge =
-  /// Maps the `'client` message to a `HMRMsg<'client>` message
-  let HMRMsgMapping = Program.UserMsg